fix(arrow/cdata): move headers into parent Go package
